+++ packages/form/src/index.tsx
@@ -33,6 +33,7 @@ import type { ModalFormProps } from './layouts/ModalForm';
 import ModalForm from './layouts/ModalForm';
 import type { DrawerFormProps } from './layouts/DrawerForm';
 import DrawerForm from './layouts/DrawerForm';
+import type { FormInstance, FormItemProps, FormProps } from './BaseForm';
 import './index.less';
 
 export type {
@@ -46,6 +47,9 @@ export type {
   QueryFilterProps,
   ProFormFieldSetProps,
   ProFormDependencyProps,
+  FormProps,
+  FormItemProps,
+  FormInstance,
 };
 
 const ProFormGroup = ProForm.Group;