+++ packages/provider/src/index.tsx
@@ -19,6 +19,7 @@ import koKR from './locale/ko_KR';
 import idID from './locale/id_ID';
 import deDE from './locale/de_DE';
 import faIR from './locale/fa_IR';
+import trTR from './locale/tr_TR';
 
 export type ProSchemaValueEnumType = {
   /** @name 演示的文案 */
@@ -136,6 +137,7 @@ const koKRIntl = createIntl('ko_KR', koKR);
 const idIDIntl = createIntl('id_ID', idID);
 const deDEIntl = createIntl('de_DE', deDE);
 const faIRIntl = createIntl('fa_IR', faIR);
+const trTRIntl = createIntl('tr_TR', trTR);
 
 const intlMap = {
   'ar-EG': arEGIntl,
@@ -154,6 +156,7 @@ const intlMap = {
   'id-ID': idIDIntl,
   'de-DE': deDEIntl,
   'fa-IR': faIRIntl,
+  'tr-TR': trTRIntl,
 };
 
 const intlMapKeys = Object.keys(intlMap);
@@ -177,6 +180,7 @@ export {
   idIDIntl,
   deDEIntl,
   faIRIntl,
+  trTRIntl,
   intlMap,
   intlMapKeys,
 };