+++ packages/table/src/components/EditableTable/demos/real-time-editing.tsx
@@ -1,4 +1,3 @@
-/** Title: 可编辑表格 */
 import React, { useState } from 'react';
 import type { ProColumns } from '@ant-design/pro-table';
 import { EditableProTable } from '@ant-design/pro-table';
@@ -94,7 +93,14 @@ export default () => {
         }}
         toolBarRender={() => {
           return [
-            <Button type="primary" key="save">
+            <Button
+              type="primary"
+              key="save"
+              onClick={() => {
+                // dataSource 就是当前数据，可以调用 api 将其保存
+                console.log(dataSource);
+              }}
+            >
               保存数据
             </Button>,
           ];