+++ packages/table/src/demos/ListToolBar/basic.tsx
@@ -1,4 +1,3 @@
-/** Title: 基本使用 */
 import React from 'react';
 import { Button } from 'antd';
 import { EllipsisOutlined, SettingOutlined, FullscreenOutlined } from '@ant-design/icons';