+++ packages/table/src/demos/pollinga.tsx
@@ -1,8 +1,9 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useState } from 'react';
 import { PlusOutlined } from '@ant-design/icons';
 import { Button } from 'antd';
-import type { ProColumns, ActionType } from '@ant-design/pro-table';
+import type { ProColumns } from '@ant-design/pro-table';
 import ProTable from '@ant-design/pro-table';
+import moment from 'moment';
 
 const valueEnum = {
   0: 'close',
@@ -86,24 +87,7 @@ const columns: ProColumns<TableListItem>[] = [
 ];
 
 export default () => {
-  const actionRef = useRef<ActionType | undefined>(undefined);
-  useEffect(() => {
-    let id = 0;
-    const loop = () => {
-      id = window.setTimeout(() => {
-        const { current } = actionRef;
-        if (current) {
-          current.reload();
-        }
-        loop();
-      }, 5000);
-    };
-    loop();
-    return () => {
-      window.clearTimeout(id);
-    };
-  }, []);
-
+  const [time, setTime] = useState(() => Date.now());
   return (
     <ProTable<TableListItem>
       columns={columns}
@@ -111,9 +95,10 @@ export default () => {
       pagination={{
         showSizeChanger: true,
       }}
-      actionRef={actionRef}
+      polling={2000}
       request={async () => {
-        await timeAwait(500);
+        await timeAwait(2000);
+        setTime(Date.now());
         return {
           data: tableListDataSource,
           success: true,
@@ -121,7 +106,7 @@ export default () => {
         };
       }}
       dateFormatter="string"
-      headerTitle="轮询"
+      headerTitle={`上次更新时间：${moment(time).format('HH:mm:ss')}`}
       toolBarRender={() => [
         <Button key="3" type="primary">
           <PlusOutlined />