+++ packages/list/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.3.3...@ant-design/pro-list@1.3.4) (2021-01-21)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.3.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.3.2...@ant-design/pro-list@1.3.3) (2021-01-21)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.3.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.3.1...@ant-design/pro-list@1.3.2) (2021-01-19)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [1.3.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.3.0...@ant-design/pro-list@1.3.1) (2021-01-18)
 
 **Note:** Version bump only for package @ant-design/pro-list